backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
Pillow==10.4.0
google-generativeai==0.8.5
numpy==1.26.4
torch==2.3.1
torchvision==0.18.1